refactor(BlackJack) : 코드 경로 수정
