+++ src/main/kotlin/blackjack/domain/participants/Participants.kt
@@ -1,5 +1,9 @@
 package blackjack.domain.participants
 
+import blackjack.domain.participants.user.Dealer
+import blackjack.domain.participants.user.Guest
+import blackjack.domain.participants.user.User
+
 data class Participants(
     val dealer: Dealer,
     val guests: List<Guest> = emptyList(),
