+++ src/main/kotlin/blackjack/domain/participants/user/Guest.kt
@@ -1,8 +1,8 @@
-package blackjack.domain.participants
+package blackjack.domain.participants.user
 
-import blackjack.domain.state.FirstTurn
-import blackjack.domain.state.Hit
 import blackjack.domain.state.State
+import blackjack.domain.state.inTurn.FirstTurn
+import blackjack.domain.state.inTurn.Hit
 
 class Guest(
     name: Name,
