+++ src/main/kotlin/blackjack/domain/participants/user/Money.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.participants
+package blackjack.domain.participants.user
 
 @JvmInline
 value class Money(val value: Int) {
