+++ src/main/kotlin/blackjack/domain/state/State.kt
@@ -2,7 +2,7 @@ package blackjack.domain.state
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.Cards
-import blackjack.domain.participants.Dealer
+import blackjack.domain.participants.user.Dealer
 import blackjack.domain.result.Outcome
 
 interface State {
