+++ src/main/kotlin/blackjack/domain/state/endTurn/BlackJack.kt
@@ -1,8 +1,8 @@
-package blackjack.domain.state
+package blackjack.domain.state.endTurn
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.Cards
-import blackjack.domain.participants.Dealer
+import blackjack.domain.participants.user.Dealer
 import blackjack.domain.result.Outcome
 
 class BlackJack(cards: Cards = Cards()) : EndTurn(cards) {
