+++ src/main/kotlin/blackjack/domain/state/inTurn/FirstTurn.kt
@@ -1,9 +1,11 @@
-package blackjack.domain.state
+package blackjack.domain.state.inTurn
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.Cards
-import blackjack.domain.participants.Dealer
+import blackjack.domain.participants.user.Dealer
 import blackjack.domain.result.Outcome
+import blackjack.domain.state.State
+import blackjack.domain.state.endTurn.BlackJack
 
 class FirstTurn private constructor(cards: Cards) : InTurn(cards) {
     constructor() : this(Cards())
