+++ src/test/kotlin/blackjack/domain/blackjack/BlackJackSettingBuilderTest.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.blackjack
 
-import blackjack.domain.participants.Name
+import blackjack.domain.participants.user.Name
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
