+++ src/test/kotlin/blackjack/domain/state/endTurn/BustTest.kt
@@ -1,9 +1,10 @@
-package blackjack.domain.state
+package blackjack.domain.state.endTurn
 
 import blackjack.domain.card.Cards
 import blackjack.domain.state.Fixtures.CLOVER_JACK
 import blackjack.domain.state.Fixtures.CLOVER_KING
 import blackjack.domain.state.Fixtures.CLOVER_QUEEN
+import blackjack.domain.state.State
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
