+++ src/test/kotlin/blackjack/domain/state/inTurn/HitTest.kt
@@ -1,10 +1,13 @@
-package blackjack.domain.state
+package blackjack.domain.state.inTurn
 
 import blackjack.domain.state.Fixtures.CLOVER_FOUR
 import blackjack.domain.state.Fixtures.CLOVER_KING
 import blackjack.domain.state.Fixtures.CLOVER_NINE
 import blackjack.domain.state.Fixtures.CLOVER_THREE
 import blackjack.domain.state.Fixtures.CLOVER_TWO
+import blackjack.domain.state.State
+import blackjack.domain.state.endTurn.Bust
+import blackjack.domain.state.endTurn.Stay
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
